cpp-port: batch 53 - port 7 hw/ files to C++
Port to .cpp:
- hw/9pfs/9p-util-generic.cpp
- hw/audio/virtio-snd-pci.cpp
- hw/usb/libhw.cpp, dev-storage-bot.cpp, vt82c686-uhci-pci.cpp,
  xlnx-usb-subsystem.cpp
- hw/virtio/virtio-rng-pci.cpp

Add extern "C" guards to 13 headers:
- include/hw/virtio/virtio.h, virtio-bus.h, virtio-pci.h, virtio-rng.h
- include/hw/audio/virtio-snd.h
- include/hw/usb/msd.h, xlnx-usb-subsystem.h
- include/hw/isa/vt82c686.h
- include/hw/register.h
- hw/usb/desc.h, hcd-uhci.h
- hw/9pfs/9p-util.h, virtio-9p.h

C++ fixes: static_cast for void* and DEV_NVECTORS_UNSPECIFIED
sign comparison, TypeInfo designator reordering.

Note: hw/virtio/virtio-9p-pci.c not ported - 9p.h uses 'private'
as struct member (C++ keyword).

+++ hw/virtio/virtio-rng-pci.cpp
@@ -44,7 +44,7 @@ static void virtio_rng_pci_realize(VirtIOPCIProxy *vpci_dev, Error **errp)
     VirtIORngPCI *vrng = VIRTIO_RNG_PCI(vpci_dev);
     DeviceState *vdev = DEVICE(&vrng->vdev);
 
-    if (vpci_dev->nvectors == DEV_NVECTORS_UNSPECIFIED) {
+    if (vpci_dev->nvectors == static_cast<uint32_t>(DEV_NVECTORS_UNSPECIFIED)) {
         vpci_dev->nvectors = 2;
     }
